build-aux/dist-vendor.sh: Add error proper checking
For automated release building the "dist" step should fail when any
command in it fails. Since we can't rely on bash features, we need to
adjust the script to fail on error.
Make it clear that this is not a bash script and rewrite it to fail
when any command fails.